Fix: #195 Split test_listing_flags.py into two independent files
